move new display setting code to rasterizer_settings
- misc code updates
